refactor: change input arg of function to unsigned int
Armstrong numbers cannot be negative. Enforce this constraint in the function
signature of is_armstrong_number by using an unsigned integer

+++ exercises/practice/armstrong-numbers/armstrong-numbers.v
@@ -1,4 +1,4 @@
 module main
 
-pub fn is_armstrong_number(number i64) bool {
+pub fn is_armstrong_number(number u32) bool {
 }
